+++ Source/GameLoop/Game.cpp
@@ -36,8 +36,6 @@ void Game::Start() {
 	ui->CrateTempFontFromTTF("Fonts/arial.ttf", 20, "arial20");
 	ui->CrateTempFontFromTTF("Fonts/arial.ttf", 40, "arial40");
 
-	Logger::SetUp("",LogOutput::Console, LogOutput::Console);
-
 	renderer->FLatRenderCopySetUp();
 }
 
@@ -55,20 +53,15 @@ void Game::Input() {
 	while (SDL_PollEvent(&event)) {
 		ui->ManageInput(event);
 		Exit();
-
-		if (event.type == SDL_KEYUP && event.key.keysym.scancode == SDL_SCANCODE_L) {
-			MT::Timer::Tic();
-			Logger::Log("Loging someting " + std::to_string(RandInt(0, 100)), LogType::Info);
-			std::println("Time: {}us", MT::Timer::Tac<std::chrono::microseconds>());
-			//std::println("Some text {}",0);
-
-		}
 	}
 	Global::inputDelay++;
 }
 
 void Game::Render() {
+	MT::Texture* tex = TexMan::GetTex("tree1");
 	renderer->ClearFrame(255, 255, 255);
+	renderer->RenderShape({ 10,10,100,100 }, tex, { 0,0,0,30 });
+	renderer->RenderShapeUPR({ 10,300,100,100 }, tex, { 0,0,0,30 });
 	ui->Render();
 	renderer->Present();
 }
@@ -90,5 +83,4 @@ Game::~Game() {
 	renderer->Clear();
 	ui->ClearAll();
 	SDL_Quit();
-	Logger::Close();
 }